% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA2csv.r
\name{CIA2csv}
\alias{CIA2csv}
\title{Output the result of Category level intensity analysis as csv.}
\usage{
CIA2csv(CIA.output, path)
}
\arguments{
\item{CIA.output}{Output list generated by \code{CIA} function.}

\item{path}{A charachter variable showing the path to where the user wants to store the csv output.}
}
\value{
The output is a CSV file.
}
\description{
Output the result of Category level intensity analysis as csv.
}
\details{
Gets the output of \code{CIA} function and the path variable and generate a csv report called "CategoryLevelIntensityAnalysis.csv". The output will be stored in "CSVOutput" directory in path direction.
}
