% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.analysis-package.r
\docType{package}
\name{intensity.analysis}
\alias{intensity.analysis}
\alias{intensity.analysis-package}
\title{Intensity Analysis on categorical maps of the same geographical extent and resolution.}
\description{
This package provides the methodology to analyz maps of categorical variables from several points in time for a single site condidering cross-tabulation matrices, where each matrix summerizes the change in each time interval. There are three levels of analysis namely: interval, category and transition. Each level tests for tests for size and stationarity of patterns across time intervals.
}
\details{
intensity.analysis.
}
\author{
Robert Gilmore Pontius Jr. <rpontius@clarku.edu>, Sam Khallaghi <SKhallaghi@clarku.edu>
}
