% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicrosstab.r
\name{multicrosstab}
\alias{multicrosstab}
\title{Generate crosstabulation table(s)}
\usage{
multicrosstab(raster.layers, time.points, categories)
}
\arguments{
\item{raster.layers}{List of raster objects.}

\item{time.points}{a charachter vector showing the time point of each raster layer in chronological order.}

\item{categories}{A charachter vector showing the categories in the map. Order of categories decided bases on the equivalent IDs in the raster attribute table.}
}
\value{
The output is a list of crosstabulation table(s). The length of the list is equal to the number of intervals.
}
\description{
Generate crosstabulation table(s)
}
\details{
Read categorical raster objects, checks their dimentionality, resolution and extent are a perfect match. Then the function returns a list consist of a crosstabulation table for each consequtive pair of raster layers.
}
