% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqpar.r
\name{reqpar}
\alias{reqpar}
\title{Calculate the required parameters for intensity analysis}
\usage{
reqpar(time.points)
}
\arguments{
\item{time.points}{a charachter vector showing the time point of each raster layer in chronological order.}
}
\value{
The output is a list inclusing: a one element vector showing the number of intervals, a charachter vector showing the initial times for all intervals, a charachter vector showing the subsequent times for all intervals, an integer vector showing the duration of each interval and a charachter vector showing the label of each interval.
}
\description{
Calculate the required parameters for intensity analysis
}
\details{
\code{reqpar} function gets the time point and return a list containing required variables used in intensity analysis.
}
