% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIA2csv.r
\name{TIA2csv}
\alias{TIA2csv}
\title{Output the result of transition level intensity analysis as csv.}
\usage{
TIA2csv(TIA.output, path)
}
\arguments{
\item{TIA.output}{Output list generated by \code{TIA} function.}

\item{path}{A charachter variable showing the path to where the user wants to store the csv output.}
}
\value{
The output is a CSV file.
}
\description{
Output the result of transition level intensity analysis as csv.
}
\details{
Gets the output of \code{TIA} function and the path variable and generate a csv report called "TransitionLevelIntensityAnalysis.csv". The output will be stored in "CSVOutput" directory in the path direction.
}
